+++ application/flicket/views/view_ticket.py
@@ -140,10 +140,17 @@ def ticket_view(ticket_id, page=1):
 
     title = gettext('View Ticket')
 
+    # display or not category change link
+    change_category = app.config['change_category']
+    if change_category and app.config['change_category_only_admin_or_super_user']:
+        if not g.user.is_admin and not g.user.is_super_user:
+            change_category = False
+
     return render_template('flicket_view.html',
                            title=title,
                            ticket=ticket,
                            form=form,
                            subscribers_form=subscribers_form,
                            replies=replies,
+                           change_category=change_category,
                            page=page)